ch09/requirements.txt
aiofiles==0.8.0
anyio==3.6.1
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
backports.zoneinfo==0.2.1
caio==0.9.6
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
ecdsa==0.18.0
fastapi==0.78.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
motor==3.0.0
odmantic==0.5.0
orjson==3.7.7
packaging==21.3
passlib==1.7.4
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.1
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.8
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.19.1
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
uvicorn==0.18.2
